+++ source/extensions/isaacsim.app.setup/isaacsim/app/setup/extension.py
@@ -271,9 +271,7 @@ class CreateSetupExtension(omni.ext.IExt):
                 MenuItemDescription(name=self.UI_SELECTOR_MENU_PATH, onclick_fn=lambda *_: self._show_selector()),
                 add_layout_menu_entry("Default", "default", carb.input.KeyboardInput.KEY_1),
                 add_layout_menu_entry("Visual Scripting", "visualScripting", carb.input.KeyboardInput.KEY_4),
-                add_layout_menu_entry(
-                    "Replicator ", "sdg", carb.input.KeyboardInput.KEY_5
-                ),  # the space after the name is intentional, otherwise it gets confused about the replicator in the tools menu
+                add_layout_menu_entry("Replicator", "sdg", carb.input.KeyboardInput.KEY_5),
                 add_layout_menu_entry("Quick Save", quick_save, carb.input.KeyboardInput.KEY_7),
                 add_layout_menu_entry("Quick Load", quick_load, carb.input.KeyboardInput.KEY_8),
             ]